Module unidraw::draw

Modules

path

shape

Structs

Anchor

Anchor refers to a vertex of a shape.

Cell

Represents a cell at a point in the cartesian plane (of the terminal ofc).

Edge

Edge refers to any of straight line formed between any two Anchors of the shape::Shape or the Terminal.

EdgeConstraint

Its synonymous to a Spring, connects any two opposite facing edges.

Point

Represents a point in the cartesian plane (of the terminal ofc).

Enums

DiagonalDirection

Measurement

Measurement is used to define dimensions of shape::Shape

StraightDirection

Type Definitions

AnchorDirection

AnchorDirection refers to direction of anchor towards the shape::Shape.

EdgeDirection

EdgeDirection refers to direction of Edge outwards the shape::Shape.